pyproject.toml
[project]
name = "serve"
version = "0.1.0"
description = "Serve HTML/Markdown files with live reload and Mermaid support"
readme = "README.md"
authors = [
    { name = "ericbryant24", email = "ericbryant24@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.9",
    "watchfiles>=1.0",
    "markdown-it-py>=3.0",
    "pygments>=2.17",
]

[project.scripts]
serve = "serve.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
